Move type helpers to formDefinitionTypes
