+++ forms-shared/src/definitions/form-definitions-helpers.ts
@@ -1,32 +1,5 @@
-import {
-  FormDefinition,
-  FormDefinitionEmail,
-  formDefinitions,
-  FormDefinitionSlovenskoSk,
-  FormDefinitionSlovenskoSkGeneric,
-  FormDefinitionSlovenskoSkTax,
-  FormDefinitionType,
-} from './form-definitions'
-
-export const isSlovenskoSkGenericFormDefinition = (
-  formDefinition: FormDefinition,
-): formDefinition is FormDefinitionSlovenskoSkGeneric =>
-  formDefinition.type === FormDefinitionType.SlovenskoSkGeneric
-
-export const isSlovenskoSkTaxFormDefinition = (
-  formDefinition: FormDefinition,
-): formDefinition is FormDefinitionSlovenskoSkTax =>
-  formDefinition.type === FormDefinitionType.SlovenskoSkTax
-
-export const isSlovenskoSkFormDefinition = (
-  formDefinition: FormDefinition,
-): formDefinition is FormDefinitionSlovenskoSk =>
-  isSlovenskoSkGenericFormDefinition(formDefinition) ||
-  isSlovenskoSkTaxFormDefinition(formDefinition)
-
-export const isEmailFormDefinition = (
-  formDefinition: FormDefinition,
-): formDefinition is FormDefinitionEmail => formDefinition.type === FormDefinitionType.Email
+import { formDefinitions } from './form-definitions'
+import { FormDefinition } from './formDefinitionTypes'
 
 export const getFormDefinitionBySlug = (slug: string): FormDefinition | null => {
   const formDefinition = formDefinitions.find((formDefinition) => formDefinition.slug === slug)
