+++ forms-shared/src/definitions/formDefinitionTypes.ts
@@ -47,3 +47,23 @@ export type FormDefinition =
   | FormDefinitionSlovenskoSkGeneric
   | FormDefinitionSlovenskoSkTax
   | FormDefinitionEmail
+
+export const isSlovenskoSkGenericFormDefinition = (
+  formDefinition: FormDefinition,
+): formDefinition is FormDefinitionSlovenskoSkGeneric =>
+  formDefinition.type === FormDefinitionType.SlovenskoSkGeneric
+
+export const isSlovenskoSkTaxFormDefinition = (
+  formDefinition: FormDefinition,
+): formDefinition is FormDefinitionSlovenskoSkTax =>
+  formDefinition.type === FormDefinitionType.SlovenskoSkTax
+
+export const isSlovenskoSkFormDefinition = (
+  formDefinition: FormDefinition,
+): formDefinition is FormDefinitionSlovenskoSk =>
+  isSlovenskoSkGenericFormDefinition(formDefinition) ||
+  isSlovenskoSkTaxFormDefinition(formDefinition)
+
+export const isEmailFormDefinition = (
+  formDefinition: FormDefinition,
+): formDefinition is FormDefinitionEmail => formDefinition.type === FormDefinitionType.Email
